pyproject.toml
[project]
name = "clinical-ml-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "faiss-cpu>=1.13.2",
    "google-genai>=1.66.0",
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.1",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.2",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]  # allow long lines in prompts/docs if needed

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = [".", "src"]
